==> kaggle_titanic/0 - generalDirectives.ipynb <==
## Первичный взгляд на данные:


* определить состав типов фич
* Which features may contain errors or typos
* Which features contain blank, null or empty values?
* What is the distribution of feature values across the samples?
* определиться, какие фичи можно дополнительно получить из сложных фич
* рассмотреть вариант генерации на основе численных значений диапазонов, назвав их цифрой (например, возраст в диапазон возрастов)
* посмотреть на выбросы

dataTrain.info()
dataTrain.describe()
dataTrain.describe(include=['O'])

* посмотреть на связь каждой переменной с целевой (может быть решить удалить её из модели)

## Analyze by visualizing data
* построрить наглядные графики. Круто, да?

На выходе - иметь идеи по модификации признаков

## Модифицировать данные (параллельно на 2-х выборках)
* выполнить идеи по изменению признаков

* удалить лишние
* сгенерить новые (features extracting)
* заполнить пропущенные
* все категориальные в числовые
* в итоге надо, чтобы получились только цифры на входе модели


Использование обобщенной выборки (train + test)
* медианы и средние для восстановления пропущенных значений можно оценивать по обоим выборкам;
* при кодировании признаков можно пробовать, чтобы кодировались только те, которые есть в обоих выборках. Остальные - "другие".

dataTrain.isnull().values.any() | dataTest.isnull().values.any()
