==> kaggle_titanic/2.2 - modifyData+fe.ipynb <==
## Modify Data accrording previous conclusions

%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np

import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.externals import joblib
from sklearn.preprocessing import StandardScaler, MinMaxScaler

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')
import supportFunctions as sf
import featureEngineringFunctions as fef

dirPath = '2.1-featuresPack/'


## Conclusions from previous notebook:

* drop features:
 * passengerId
 * cabin (after extract the deck)
 * ticket
 * embarked
* missing data to recover:
 * <s>age</s>
* convert categorial to:
 * <s>sex - to int and to several columns</s>
* new features from:
 * <s>cabin (extract decks)</s>
 * <s>name (extract title)</s>
 * <s>age bands</s>
 * <s>fare </s>normalize and (?)<s> make bands</s>
 * <s>sibsp & parch - may be generated isAlone hasChildren/Parents hasHasb/Wife/</s>
 
 Finally convert to smaller memory size

#dataRawTrain, dataRawTest, _ = sf.load_raw_data()
dataTrain, dataTest, _ = sf.load_raw_data()
dataAll = dataTrain.append(dataTest, ignore_index=True)
#dataAll.tail()

trainVStest = len(dataTrain)
#test_copy = dataTest.copy()

dataAll.isnull().sum()

## Id + Survived

# save base attrs
fef.save_to_file(dataAll['PassengerId'], 'PassengerId', dirPath)
fef.save_to_file(dataAll['Survived'], 'Survived', dirPath)

### Class

fef.save_to_file (dataAll['Pclass'], 'Pclass_LE', dirPath)

### SEX

#no missing values

# encode SEX feature

#save as OHE
featureSex, col_ = fef.fe_cat_ohe(dataAll, 'Sex')
fef.save_to_file (featureSex, 'Sex_OHE', dirPath)

#save as LE
dataAll['Sex'] = dataAll['Sex'].map({'female': 1, 'male': 0}).astype(int)
featureSex, col_ = fef.fe_cat_le(dataAll, 'Sex')

fef.save_to_file (featureSex, 'Sex_LE', dirPath)

dataAll.head()

### Missing data recovery, AGE
Strategy = median Age for Pclass=1 and Gender=0, Pclass=1 and Gender=1, and so on...

sf.check_missing_data(dataTrain, dataTest, ['Age'])
#dataTrain.loc[dataTrain['Age'].isnull() == True]

guess_ages = np.zeros((2,3))

for dataset in [dataAll[:trainVStest], dataAll[trainVStest:]]:
    for i in range(0, 2):
        for j in range(0, 3):
            guess_df = dataset[(dataset['Sex'] == i) & (dataset['Pclass'] == j+1)]['Age'].dropna()
            age_guess = guess_df.median()

            # Convert random age float to nearest .5 age
            guess_ages[i,j] = int(age_guess/0.5 + 0.5) * 0.5
            
    for i in range(0, 2):
        for j in range(0, 3):
            dataset.loc[(dataset.Age.isnull()) & (dataset.Sex == i) & (dataset.Pclass == j+1), 'Age'] = guess_ages[i,j]

#dataTrain.head()
#dataAll = dataTrain.append(dataTest)
sf.check_missing_data(dataAll[:trainVStest], dataAll[trainVStest:], ['Age'])

#fef.save_to_file (dataTrain['Age'], dataTest['Age'], dirPath, 'AGE_cont')
fef.save_to_file (dataAll['Age'], 'AGE_cont', dirPath)

### Extract title from name

dataAll['Title'] = dataAll.Name.str.extract(' ([A-Za-z]+)\.', expand=False)
    
dataAll.Title.value_counts()

dataAll['Title'] = dataAll['Title'].replace(['Countess','Capt', 'Col', 
                                             'Don', 'Dr', 'Major', 'Rev', 'Sir', 'Jonkheer', 'Dona'], 'Rare')

dataAll['Title'] = dataAll['Title'].replace(['Mlle','Ms', 'Lady'], 'Miss')
dataAll['Title'] = dataAll['Title'].replace('Mme', 'Mrs')
    
dataAll[:trainVStest][['Title', 'Survived']].groupby(['Title'], as_index=False).mean()

#save as OHE
featureTitle, col_ = fef.fe_cat_ohe(dataAll, 'Title')
fef.save_to_file (featureTitle, 'Title_OHE', dirPath)

title_mapping = {"Mr": 1, "Miss": 2, "Mrs": 3, "Master": 4, "Rare": 5}
dataAll['Title'] = dataAll['Title'].map(title_mapping)

#save as LE
fef.save_to_file (dataAll['Title'], 'Title_LE', dirPath)

### Extract AgeBands from Age

dataAll['AgeBin'] = pd.qcut(dataAll['Age'], 5)
dataAll[:trainVStest][['AgeBin', 'Survived']].groupby(['AgeBin'], as_index=False).mean().sort_values(by='AgeBin', ascending=True)

featureAgeBin, col_ = fef.fe_cat_le(dataAll, 'AgeBin')
fef.save_to_file (featureAgeBin, 'AgeBin_LE', dirPath)
dataAll.drop('AgeBin', axis=1, inplace=True)

#dataAll.head(5)


### sibsp & parch new features generation

dataAll['FamilySize_LE'] = dataAll['SibSp'] + dataAll['Parch'] + 1
fef.save_to_file (dataAll['FamilySize_LE'], 'FamilySize_LE', dirPath)

fef.save_to_file (dataAll['SibSp'], 'SibSp_LE', dirPath)
fef.save_to_file (dataAll['Parch'], 'Parch_LE', dirPath)

dataAll[:trainVStest][['FamilySize_LE', 'Survived']].groupby(['FamilySize_LE'], as_index=False).mean().sort_values(by='Survived', ascending=False)

dataAll['IsAlone_LE'] = 0
dataAll.loc[dataAll['FamilySize_LE'] == 1, 'IsAlone_LE'] = 1
dataAll['HasWifeHasb_LE'] = 0
dataAll.loc[dataAll['SibSp'] > 0 , 'HasWifeHasb'] = 1
dataAll['HasBloodRelatives_LE'] = 0
dataAll.loc[dataAll['Parch'] > 0 , 'HasWifeHasb'] = 1

fef.save_to_file (dataAll['IsAlone_LE'], 'IsAlone_LE', dirPath)
fef.save_to_file (dataAll['HasWifeHasb_LE'], 'HasWifeHasb_LE', dirPath)
fef.save_to_file (dataAll['HasBloodRelatives_LE'], 'HasBloodRelatives_LE', dirPath)

dataAll[:trainVStest][['IsAlone_LE', 'Survived']].groupby(['IsAlone_LE'], as_index=False).mean()

#dataTrain

### Syntetical union of features

dataAll['Age*Class'] = (dataAll.Age * dataAll.Pclass).astype(int)

fef.save_to_file (dataAll['Age*Class'], 'Age-Class_cont', dirPath)
    
dataAll.loc[:, ['Age*Class', 'Age', 'Pclass']].head(10)

### Fare levels

dataAll.isnull().sum()

#for test only
def get_median_fare_by_item (pClass, i):
    return dataAll.query('Pclass == {}'.format(pClass))['Fare'].mean()
    
for i in range(len (dataAll)):
    if pd.isnull(dataAll.loc[i,'Fare']):
        dataAll.loc[i,'Fare'] = get_median_fare_by_item (dataAll.loc[i,'Pclass'], i)

dataAll['FareBin'] = pd.qcut(dataAll['Fare'], 5)
dataAll[['FareBin', 'Survived']].groupby(['FareBin'], as_index=False).mean().sort_values(by='FareBin', ascending=True)

#for dataset in [dataTrain, dataTest]:
#    dataset['Fare_lev'] = 0
#    dataset['Fare_lev'].loc[dataset['Fare'] <= 7.85] = 0
#    dataset['Fare_lev'].loc[(dataset['Fare'] > 7.85) & (dataset['Fare'] <= 10.5)] = 1
#    dataset['Fare_lev'].loc[(dataset['Fare'] > 10.5) & (dataset['Fare'] <= 21.67)] = 2
#    dataset['Fare_lev'].loc[(dataset['Fare'] > 21.67) & (dataset['Fare'] <= 39.7)] = 3
#    dataset['Fare_lev'].loc[dataset['Fare'] > 39.7] = 4

#fef.save_to_file (dataTrain['Fare_lev'], dataTest['Fare_lev'], dirPath, 'Fare_lev') 

#dataTrain = dataTrain.drop(['FareBin'], axis=1)

#fef.save_to_file (dataTrain['Fare'], dataTest['Fare'], dirPath, 'Fare_cnt')

#for dataset in [dataTrain, dataTest]:
#    dataset.drop(['Fare'], axis=1)

fef.save_to_file (dataAll['Fare'], 'Fare_cont', dirPath)

featureAgeBin, col_ = fef.fe_cat_le(dataAll, 'FareBin')
fef.save_to_file (featureAgeBin, 'FareBin_LE', dirPath)
dataAll.drop('FareBin', axis=1, inplace=True)




### Cabin to decks

#decks = {"U": 0, "A": 1, "B": 2, "C": 3, "D": 4, "E": 5, "F": 6, "G": 7, "T": 0}

dataAll['Deck'] = 0
dataAll['Deck'] = dataAll['Cabin'].fillna('U').apply(lambda c: c[0])
print (dataAll.Deck.value_counts())

dataAll['Deck'] = dataAll['Deck'].replace('T', 'U')

print (dataAll.Deck.value_counts())

#save as OHE
featureDeck, col_ = fef.fe_cat_ohe(dataAll, 'Deck')
fef.save_to_file (featureDeck, 'Deck_OHE', dirPath)

#save as LE
featureDeck, col_ = fef.fe_cat_le(dataAll, 'Deck')
fef.save_to_file (featureDeck, 'Deck_LE', dirPath)

## Stolen features

### FamilySurvival

# This feature is from S.Xu, https://www.kaggle.com/shunjiangxu/blood-is-thicker-than-water-friendship-forever
dataAll['LastName'] = dataAll['Name'].apply(lambda x: str.split(x, ",")[0])
#total['Fare'].fillna(total['Fare'].mean(), inplace=True)

default_survival_rate = 0.5
dataAll['FamilySurvival'] = default_survival_rate

for grp, grp_df in dataAll[['Survived','Name', 'LastName', 'Fare', 'Ticket', 'PassengerId',
                           'SibSp', 'Parch', 'Age', 'Cabin']].groupby(['LastName', 'Fare']):
    
    if (len(grp_df) != 1):
        # A Family group is found.
        for ind, row in grp_df.iterrows():
            smax = grp_df.drop(ind)['Survived'].max()
            smin = grp_df.drop(ind)['Survived'].min()
            passID = row['PassengerId']
            if (smax == 1.0):
                dataAll.loc[dataAll['PassengerId'] == passID, 'FamilySurvival'] = 1
            elif (smin==0.0):
                dataAll.loc[dataAll['PassengerId'] == passID, 'FamilySurvival'] = 0

print("Number of passengers with family survival information:", 
      dataAll.loc[dataAll['FamilySurvival']!=0.5].shape[0])
dataAll['FamilySurvival'].value_counts()


for _, grp_df in dataAll.groupby('Ticket'):
    if (len(grp_df) != 1):
        for ind, row in grp_df.iterrows():
            if (row['FamilySurvival'] == 0) | (row['FamilySurvival']== 0.5):
                smax = grp_df.drop(ind)['Survived'].max()
                smin = grp_df.drop(ind)['Survived'].min()
                passID = row['PassengerId']
                if (smax == 1.0):
                    dataAll.loc[dataAll['PassengerId'] == passID, 'FamilySurvival'] = 1
                elif (smin==0.0):
                    dataAll.loc[dataAll['PassengerId'] == passID, 'FamilySurvival'] = 0
                        
print("Number of passenger with family/group survival information: " 
      +str(dataAll[dataAll['FamilySurvival']!=0.5].shape[0]))
fef.save_to_file (dataAll['FamilySurvival'], 'FamilySurvival_bin', dirPath)
dataAll['FamilySurvival'].value_counts()

dataAll.head()

for _, grp_df in total.groupby('Ticket'):
    if (len(grp_df) != 1):
        for ind, row in grp_df.iterrows():
            if (row['Family_Survival'] == 0) | (row['Family_Survival']== 0.5):
                smax = grp_df.drop(ind)['Survived'].max()
                smin = grp_df.drop(ind)['Survived'].min()
                passID = row['PassengerId']
                if (smax == 1.0):
                    total.loc[total['PassengerId'] == passID, 'Family_Survival'] = 1
                elif (smin==0.0):
                    total.loc[total['PassengerId'] == passID, 'Family_Survival'] = 0
                        
print("Number of passenger with family/group survival information: " 
      +str(total[total['Family_Survival']!=0.5].shape[0]))

### Drop features

#dataAll.head()

#no sense in drop - all features are in files


### Final checkers missing values

dataAll.isnull().sum()
#TODO неплохо было бы проверять на уникальные значения в тесте и трейне - чтобы бились

## what can be done more

- можно восстановить палубу исходя из класса и ценника билета
1 класс - A-E, самые дорогие в B
2 и 3 класс - D-G, в D - лучшие

- Syntetical union of features

#features_from_other = ['Survived','Pclass','Sex','Family_Size','Family_Survival','Fare_Bin','Age_Bin']
featureListForTest = [
#'PassengerId',
#'Survived',
'AGE_cont',
'AgeBin_LE',
'Age-Class_cont',
'Deck_LE',
'Deck_OHE',
'FamilySize_LE',
'Fare_cont',
'FareBin_LE',
'HasBloodRelatives_LE',
'HasWifeHasb_LE',
'IsAlone_LE',
'Parch_LE',
'Pclass_LE',
'Sex_LE',
'Sex_OHE',
'SibSp_LE',
'Title_LE',
'Title_OHE',
'FamilySurvival_bin'
]
dataAll2 = fef.create_dataset_from_features (featureListForTest, '2.1-featuresPack/')
dataAll2#.head()
#dataAll2[trainVStest:].head(500)
