==> kaggle_titanic/2.7 - Memory optimizations.ipynb <==
## Memory optimizations

%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.externals import joblib
from sklearn.preprocessing import StandardScaler, MinMaxScaler

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')
import supportFunctions as sf


from sklearn import preprocessing

@ надо грузить из других файлов
dataRawTrain, dataRawTest, _ = sf.load_raw_data()
dataTrain, dataTest, _ = sf.load_raw_data()
#dataTrain.head()

### Types optimizations

#before
dataTrain.info()
dataTest.info()
#dataTrain.memory_usage()
#dataTest.memory_usage()

for dataset in [dataTrain, dataTest, dataTrainScaled, dataTestScaled]:
    for column in list(dataset.columns):
        if dataset[column].dtype == 'int64':
            dataset[column] = dataset[column].astype(int)
            continue
        if dataset[column].dtype == 'float64':
            dataset[column] = dataset[column].astype(np.dtype('Float16'))
            continue
        if dataset[column].dtype == 'int32':
            dataset[column] = dataset[column].astype(np.dtype('uint8'))
            continue

dataTrain.info()
dataTest.info()

### Save data to dump

#TODO дату в файле
@comment_if_you_are_ready
_ = joblib.dump([dataTrain, dataTrainTarget, dataTest, dataTrainScaled, dataTestScaled], 
            open('data_modified/dataModified.pkl', 'wb'), 9)
