==> kaggle_titanic/3 - checkClfModelsRougly.ipynb <==
%load_ext autoreload
%autoreload 2
import pandas as pd
from sklearn.externals import joblib
from sklearn.model_selection import StratifiedKFold
from sklearn.preprocessing import StandardScaler

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')

import CheckClassificationModels as ccm
import supportFunctions as sf
import featureEngineringFunctions as fef

dataRawTrain, dataRawTest, _ = sf.load_raw_data()

Survived_train 
AGE_cnt_train
Age_lev_train
Age-Class_cnt_train
Deck_LE_train
Deck_OHE_train
FamilySize_cnt_train
Fare_cnt_train
Fare_lev_train
HasBloodRelatives_train
HasWifeHasb_train
IsAlone_train
PassengerId_train
Pclass_cnt_train
Pclass_OHE_train
SEX_LE_train
SEX_OHE_train
Title_LE_train
Title_OHE_train

featureListLE = [
'AGE_cnt_train',
'Age-Class_cnt_train',
'Deck_LE_train',
'FamilySize_cnt_train',
'Fare_cnt_train',
'HasBloodRelatives_train',
'HasWifeHasb_train',
'IsAlone_train',
'Pclass_cnt_train',
'SEX_LE_train',
'Title_LE_train'
]
dataTrainLE = fef.create_dataset_from_features (featureListLE, '2.1-featuresPack/')
#dataTrainLE

featureListLElev = [
'AGE_lev_train',
'Age-Class_cnt_train',
'Deck_LE_train',
'FamilySize_cnt_train',
'Fare_lev_train',
'HasBloodRelatives_train',
'HasWifeHasb_train',
'IsAlone_train',
'Pclass_cnt_train',
'SEX_LE_train',
'Title_LE_train'
]
dataTrainLElev = fef.create_dataset_from_features (featureListLElev, '2.1-featuresPack/')
#dataTrainLE

featureListOHE = [
'AGE_cnt_train',
'Age-Class_cnt_train',
'Deck_OHE_train',
'FamilySize_cnt_train',
'Fare_cnt_train',
'HasBloodRelatives_train',
'HasWifeHasb_train',
'IsAlone_train',
'Pclass_cnt_train',
'SEX_OHE_train',
'Title_OHE_train'
]
dataTrainOHE = fef.create_dataset_from_features (featureListOHE, '2.1-featuresPack/')
#dataTrainOHE

featureListOHElev = [
'AGE_lev_train',
'Age-Class_cnt_train',
'Deck_OHE_train',
'FamilySize_cnt_train',
'Fare_lev_train',
'HasBloodRelatives_train',
'HasWifeHasb_train',
'IsAlone_train',
'Pclass_cnt_train',
'SEX_OHE_train',
'Title_OHE_train'
]
dataTrainOHElev = fef.create_dataset_from_features (featureListOHElev, '2.1-featuresPack/')
dataTrainOHElev

datasetPack = [dataTrainLE, dataTrainOHE, dataTrainLElev, dataTrainOHElev]

## задача этого ноута - построить кроссвалидацию + все начальные модели

## Crossvalidation

#check if balanced or not
print ('check if balanced or not: {}'.format(float(sum(dataRawTrain.Survived))/len(dataRawTrain.Survived)))

#http://scikit-learn.org/stable/modules/cross_validation.html#cross-validation-iterators
# checklist CV strategy
# данные приходят из i.i.d распределения?
# данные имеют временную зависимоть?
# данные имеют зависимость от группы? например, строки можно объединить в группы, где целевая переменная генерится по своему закону по каждой группе?
# может выделить искуственную группу от признака, от которого есть большая корреляция?
# есть ли смежные данные?

# TODO metrics should be choosen optimally

RS = 100
nSamples = len(dataRawTrain)
nSplits = [2,4,6,10,15,20]
scoring = 'roc_auc'

for dataset in datasetPack[:]:
    cvs = [StratifiedKFold(n_splits=i, random_state=RS, shuffle=True) for i in nSplits]
    
    datasetMods = [
        ['scaled', StandardScaler().fit_transform(dataset[:nSamples])],
        ['non-scaled', dataset[:nSamples]]
        ]

    results = ccm.global_check_clf_models (datasetMods, dataRawTrain.Survived[:nSamples], scoring, cvs, RS, n_jobs = -1)#, debugMode=1)
    ccm.postprocess_and_save_results(results, list(dataset.columns), cvs, datasetMods, path='checkClfModels/')
