==> kaggle_titanic/5.1 - modelTuningSVC.ipynb <==
from sklearn.svm import SVC
%load_ext autoreload
%autoreload 2
import pandas as pd
from sklearn.externals import joblib
from sklearn.model_selection import StratifiedKFold

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')
import GridSearchCVcustom as GSC



#load data
dataTrain, dataTarget, _ , dataTrainScaled, _, selectedFeatures = joblib.load('dumps_clean_data/dataDump.pkl')
dataTrain = dataTrain[selectedFeatures]
dataTrainScaled = dataTrainScaled[selectedFeatures]

#check available params
pipeline = GSC.get_pipeline_1step(SVC())
pipeline.named_steps 

RS = 100
n_splits = 10
nSamples = len(dataTrain)
cvShuffle = False
paramGrid = {'clf': [SVC()],
              'clf__random_state': [RS],
              #'clf__class_weight': ['balanced'],
              'clf__C': [0.01, 0.1, 0.5, 1., 1.5, 2.5, 5, 10., 50.],
              'clf__max_iter': [-1]              
             }

# non-Scaled + StratifiedKFold

cv = StratifiedKFold(n_splits=n_splits, random_state=RS, shuffle=cvShuffle)

bestScore, bestScoreParams, bestIter = GSC.GridSearchCVcustom(pipeline, paramGrid, 
                                                              dataTrain[:nSamples], dataTarget[:nSamples],
                                                              cv=cv, n_jobs = -1, pre_dispatch=2,
                                                              showSteps=True, itersToPerform ='All')
print ('bestScore = {}\nBestScoreParams = {}'.format(bestScore, bestScoreParams))

# Scaled + StratifiedKFold

cv = StratifiedKFold(n_splits=n_splits, random_state=RS, shuffle=cvShuffle)

bestScore, bestScoreParams, bestIter = GSC.GridSearchCVcustom(pipeline, paramGrid, 
                                                              dataTrainScaled[:nSamples], dataTarget[:nSamples],
                                                              cv=cv, n_jobs = -1, pre_dispatch=2,
                                                              showSteps=True, itersToPerform ='All')
print ('bestScore = {}\nBestScoreParams = {}'.format(bestScore, bestScoreParams))
