==> kaggle_titanic/5.4 - modelTuningXGB.ipynb <==
%load_ext autoreload
%autoreload 2
import pandas as pd
from sklearn.externals import joblib

from sklearn.model_selection import GridSearchCV
import xgboost as xgb

from sklearn.model_selection import StratifiedKFold

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')
import GridSearchCVcustom as GSC
import featureEngineringFunctions as fef
from sklearn.preprocessing import StandardScaler

dirPath = '2-featuresPack/'

#featuresFromOtherSolutions = ['Survived','Pclass','Sex','Family_Size','Family_Survival','Fare_Bin','Age_Bin']

featureList = [
#'PassengerId',
#'Survived',
#'AGE_cont',
'AgeBin_LE',
#'Age-Class_cont',
#'Deck_LE',
#'Deck_OHE',
'FamilySize_LE',
#'Fare_cont',
'FareBin_LE',
#'HasBloodRelatives_LE',
#'HasWifeHasb_LE',
#'IsAlone_LE',
#'Parch_LE',
'Pclass_LE',
'Sex_LE',
#'Sex_OHE',
#'SibSp_LE',
#'Title_LE',
#'Title_OHE',
'FamilySurvival_bin'
]
#features = ['Survived','Pclass','Sex','Family_Size','Family_Survival','Fare_Bin','Age_Bin']

trainVStest = 891
dataAll = fef.create_dataset_from_features (featureList, dirPath)
dataTrain = fef.create_dataset_from_features (featureList, dirPath)[:trainVStest]
dataTrainTarget =fef.create_dataset_from_features (['Survived'], dirPath)[:trainVStest]

dataTrainScaled = StandardScaler().fit(dataAll).transform(dataTrain)

#check available params
pipeline = GSC.get_pipeline_1step(xgb.XGBClassifier(n_jobs=-1))
pipeline.named_steps 

RS = 100
score = "roc_auc"
n_splits = 5
nSamples = len(dataTrain)
cvShuffle = True
paramGrid = {'clf': [xgb.XGBClassifier(n_jobs=-1)],
              'clf__random_state': [RS],
              'clf__class_weight': ['balanced'],
              'clf__n_estimators': [10, 100, 250, 500, 750, 1000],
              'clf__max_depth': [1, 2, 4],
              'clf__learning_rate': [0.01, 0.02, 0.05, 0.1, 0.5]
             }
cv = StratifiedKFold(n_splits=n_splits, random_state=RS, shuffle=cvShuffle)

# non-Scaled + StratifiedKFold

bestScore1, bestScoreParams1, bestIter1 = GSC.GridSearchCVcustom(pipeline, paramGrid, 
                                                              dataTrain, dataTrainTarget,
                                                              cv=cv, n_jobs = -1, scoring = score, pre_dispatch=2,
                                                              showSteps=True, itersToPerform ='All')
print ('bestScore = {}\nBestScoreParams = {}'.format(bestScore1, bestScoreParams1))

# Scaled + StratifiedKFold

bestScore2, bestScoreParams2, bestIter2 = GSC.GridSearchCVcustom(pipeline, paramGrid, 
                                                              dataTrainScaled, dataTrainTarget,
                                                              cv=cv, n_jobs = -1, scoring = score, pre_dispatch=2,
                                                              showSteps=True, itersToPerform ='All')
print ('bestScore = {}\nBestScoreParams = {}'.format(bestScore2, bestScoreParams2))

## Pure GS

clf = xgb.XGBClassifier(n_jobs=-1)

params = {
              'random_state': [RS],
              #'class_weight': ['balanced'],
              'n_estimators': [10, 100, 250, 500, 750, 1000],
              'max_depth': [1, 2, 4],
              'learning_rate': [0.01, 0.02, 0.05, 0.1, 0.5]
             }

# Using ROC_AUC as metric has a better result than using accuracy. 
gs = GridSearchCV(clf, param_grid= params, cv = cv, n_jobs = -1 ,scoring = score,verbose=2)
gs.fit(dataTrainScaled, dataTrainTarget)
print(gs.best_score_)
print(gs.best_estimator_)

gs.best_estimator_

cv

## save to disk

#model = bestScoreParams2['clf']
model = gs.best_estimator_
model.fit(dataTrainScaled, dataTrainTarget)

@comment_this_line_if_ready
from datetime import datetime
_ = joblib.dump([model, featureList], 
            open('modelsTrained/modelXGB--{}.pkl'.format(datetime.now().strftime("%Y-%m-%d--%H-%M")), 'wb'), 9)
