==> kaggle_titanic/8 -  predictions.ipynb <==
%load_ext autoreload
%autoreload 2
import pandas as pd
from sklearn.externals import joblib
#import xgboost as xgb
#from lightgbm import LGBMClassifier
from sklearn.preprocessing import StandardScaler

import sys
sys.path.insert(0, 'G:/work/GitHub/ml_baseline_things/functions/')
#import GridSearchCVcustom as GSC
import featureEngineringFunctions as fef
dirPath = '2-featuresPack/'

#load data
#_, _, dataTest, _, dataTestScaled = joblib.load('data_modified/dataModified.pkl')

fileName = 'modelKNN--2018-12-07--15-17.pkl'
model, selectedFeatures = joblib.load('modelsTrained/' + fileName)
selectedFeatures

trainVStest = 891

dataAll = fef.create_dataset_from_features (selectedFeatures, dirPath)
dataTest = fef.create_dataset_from_features (selectedFeatures, dirPath)[trainVStest:]
dataTestPassengerID =fef.create_dataset_from_features (['PassengerId'], dirPath)[trainVStest:]

dataTestScaled = StandardScaler().fit(dataAll).transform(dataTest)

#@
prediction = model.predict(dataTestScaled).astype(int)

submission = pd.DataFrame({
        "PassengerId": dataTestPassengerID,
        "Survived": prediction
    })
submission.to_csv('submissions/submission.csv', index=False)
