==> kaggle_titanic/9 -  predictions postprocess.ipynb <==
%load_ext autoreload
%autoreload 2
import pandas as pd
from sklearn.externals import joblib
import xgboost as xgb
from lightgbm import LGBMClassifier

#load data
_, _, dataTest, _, dataTestScaled = joblib.load('data_modified/dataModified.pkl')

#model, selectedFeatures = joblib.load('modelsTrained/' + fileName)
dataTestRaw = pd.read_csv('data_raw/test.csv')
submission = pd.read_csv('submissions/submission - pl XGBscaled 0.77990.csv')

## Postprocess 1: save all children below 15 yo depending on class

submission['Survived'].value_counts()
#dataTestRaw.query('(Pclass == 1 or Pclass == 2) and Age<17')
#dataTestRaw.query('Age<17')

children_idx = submission['PassengerId'].isin(dataTestRaw.query('Age<17').PassengerId)
submission['Survived'].loc[children_idx] = 1

submission['Survived'].value_counts()

#@
submission.to_csv('submissions/submission(postprocess1).csv', index=False)
